Addition of the region sharding upgraded files

+++ test_mysql/database/init/alter_users_tables_sharding.sql
@@ -1,1 +1,1 @@
-ALTER TABLE user change _id _id BIGINT NOT NULL;
+ALTER TABLE user change _id _id INT NOT NULL;
